+++ src/__tests__/autocapture.test.ts
@@ -368,6 +368,8 @@ describe('Autocapture system', () => {
     describe('_captureEvent', () => {
         beforeEach(() => {
             posthog.config.rageclick = true
+            // Trigger proper enabling
+            autocapture.afterDecideResponse({} as DecideResponse)
         })
 
         it('should capture rageclick', () => {
@@ -914,12 +916,44 @@ describe('Autocapture system', () => {
                 'a.test-class.test-class2.test-class3.test-class4.test-class5:nth-child="1"nth-of-type="1"href="http://test.com"attr__href="http://test.com"attr__class="test-class test-class2 test-class3 test-class4 test-class5";span:nth-child="1"nth-of-type="1"'
             )
         })
+
+        it('correctly captures text when multiple button children elements', () => {
+            const parent = document.createElement('div')
+            const button = document.createElement('button')
+            const image = document.createElement('img')
+            const textOne = document.createElement('span')
+            textOne.textContent = 'the button text'
+            const textTwo = document.createElement('span')
+            textTwo.textContent = `
+            with more
+            <!-- -->
+            info
+            `
+            parent.appendChild(button)
+            button.appendChild(image)
+            button.appendChild(textOne)
+            button.appendChild(textTwo)
+
+            const e = {
+                target: image,
+                type: 'click',
+            } as unknown as MouseEvent
+
+            autocapture['_captureEvent'](e)
+
+            expect(captureMock).toHaveBeenCalledTimes(1)
+            const props = captureMock.mock.calls[0][1]
+            const capturedButton = props['$elements'][1]
+            expect(capturedButton['tag_name']).toBe('button')
+            expect(capturedButton['$el_text']).toBe('the button text with more <!-- --> info')
+        })
     })
 
     describe('_addDomEventHandlers', () => {
         beforeEach(() => {
             document.title = 'test page'
             posthog.config.mask_all_element_attributes = false
+            autocapture.afterDecideResponse({} as DecideResponse)
         })
 
         it('should capture click events', () => {
@@ -943,7 +977,12 @@ describe('Autocapture system', () => {
             jest.spyOn(autocapture, '_addDomEventHandlers')
         })
 
-        it('should be enabled before the decide response', () => {
+        it('should not be enabled before the decide response', () => {
+            expect(autocapture.isEnabled).toBe(false)
+        })
+
+        it('should be enabled before the decide response if decide is disabled', () => {
+            posthog.config.advanced_disable_decide = true
             expect(autocapture.isEnabled).toBe(true)
         })