+++ src/__tests__/decide.js
@@ -1,6 +1,7 @@
 import { Decide } from '../decide'
 import { PostHogPersistence } from '../posthog-persistence'
 import { RequestRouter } from '../utils/request-router'
+import { expectScriptToExist, expectScriptToNotExist } from './helpers/script-utils'
 
 const expectDecodedSendRequest = (send_request, data, noCompression) => {
     const lastCall = send_request.mock.calls[send_request.mock.calls.length - 1]
@@ -19,25 +20,6 @@ const expectDecodedSendRequest = (send_request, data, noCompression) => {
     })
 }
 
-const checkScriptsForSrc = (src, negate = false) => {
-    const scripts = document.querySelectorAll('body > script')
-    let foundScript = false
-    for (let i = 0; i < scripts.length; i++) {
-        if (scripts[i].src === src) {
-            foundScript = true
-            break
-        }
-    }
-
-    if (foundScript && negate) {
-        throw new Error(`Script with src ${src} was found when it should not have been.`)
-    } else if (!foundScript && !negate) {
-        throw new Error(`Script with src ${src} was not found when it should have been.`)
-    } else {
-        return true
-    }
-}
-
 describe('Decide', () => {
     given('decide', () => new Decide(given.posthog))
     given('posthog', () => ({
@@ -51,16 +33,6 @@ describe('Decide', () => {
         _afterDecideResponse: jest.fn(),
         get_distinct_id: jest.fn().mockImplementation(() => 'distinctid'),
         _send_request: jest.fn().mockImplementation(({ callback }) => callback?.({ config: given.decideResponse })),
-        toolbar: {
-            maybeLoadToolbar: jest.fn(),
-            afterDecideResponse: jest.fn(),
-        },
-        sessionRecording: {
-            afterDecideResponse: jest.fn(),
-        },
-        autocapture: {
-            afterDecideResponse: jest.fn(),
-        },
         featureFlags: {
             receivedFeatureFlags: jest.fn(),
             setReloadingPaused: jest.fn(),
@@ -197,8 +169,6 @@ describe('Decide', () => {
             given('decideResponse', () => ({}))
             given.subject()
 
-            expect(given.posthog.sessionRecording.afterDecideResponse).toHaveBeenCalledWith(given.decideResponse)
-            expect(given.posthog.toolbar.afterDecideResponse).toHaveBeenCalledWith(given.decideResponse)
             expect(given.posthog.featureFlags.receivedFeatureFlags).toHaveBeenCalledWith(given.decideResponse, false)
             expect(given.posthog._afterDecideResponse).toHaveBeenCalledWith(given.decideResponse)
             expect(given.posthog.autocapture.afterDecideResponse).toHaveBeenCalledWith(given.decideResponse)
@@ -228,10 +198,7 @@ describe('Decide', () => {
 
             given.subject()
 
-            expect(given.posthog.autocapture.afterDecideResponse).toHaveBeenCalledWith(given.decideResponse)
-            expect(given.posthog.sessionRecording.afterDecideResponse).toHaveBeenCalledWith(given.decideResponse)
-            expect(given.posthog.toolbar.afterDecideResponse).toHaveBeenCalledWith(given.decideResponse)
-
+            expect(given.posthog._afterDecideResponse).toHaveBeenCalledWith(given.decideResponse)
             expect(given.posthog.featureFlags.receivedFeatureFlags).not.toHaveBeenCalled()
         })
 
@@ -248,10 +215,7 @@ describe('Decide', () => {
 
             given.subject()
 
-            expect(given.posthog.autocapture.afterDecideResponse).toHaveBeenCalledWith(given.decideResponse)
-            expect(given.posthog.sessionRecording.afterDecideResponse).toHaveBeenCalledWith(given.decideResponse)
-            expect(given.posthog.toolbar.afterDecideResponse).toHaveBeenCalledWith(given.decideResponse)
-
+            expect(given.posthog._afterDecideResponse).toHaveBeenCalledWith(given.decideResponse)
             expect(given.posthog.featureFlags.receivedFeatureFlags).not.toHaveBeenCalled()
         })
 
@@ -259,7 +223,7 @@ describe('Decide', () => {
             given('config', () => ({ api_host: 'https://test.com', opt_in_site_apps: true, persistence: 'memory' }))
             given('decideResponse', () => ({ siteApps: [{ id: 1, url: '/site_app/1/tokentoken/hash/' }] }))
             given.subject()
-            expect(checkScriptsForSrc('https://test.com/site_app/1/tokentoken/hash/')).toBe(true)
+            expectScriptToExist('https://test.com/site_app/1/tokentoken/hash/')
         })
 
         it('does not run site apps code if not opted in', () => {
@@ -272,56 +236,7 @@ describe('Decide', () => {
                 // throwing only in tests, just an error in production
                 'Unexpected console.error: [PostHog.js],PostHog site apps are disabled. Enable the "opt_in_site_apps" config to proceed.'
             )
-            expect(checkScriptsForSrc('https://test.com/site_app/1/tokentoken/hash/', true)).toBe(true)
-        })
-
-        it('Make sure surveys are not loaded when decide response says no', () => {
-            given('decideResponse', () => ({
-                featureFlags: { 'test-flag': true },
-                surveys: false,
-            }))
-            given('config', () => ({
-                api_host: 'https://test.com',
-                token: 'testtoken',
-                persistence: 'memory',
-            }))
-
-            given.subject()
-            // Make sure the script is not loaded
-            expect(checkScriptsForSrc('https://test.com/static/surveys.js', true)).toBe(true)
-        })
-
-        it('Make sure surveys are loaded when decide response says so', () => {
-            given('decideResponse', () => ({
-                featureFlags: { 'test-flag': true },
-                surveys: true,
-            }))
-            given('config', () => ({
-                api_host: 'https://test.com',
-                token: 'testtoken',
-                persistence: 'memory',
-            }))
-
-            given.subject()
-            // Make sure the script is loaded
-            expect(checkScriptsForSrc('https://test.com/static/surveys.js')).toBe(true)
-        })
-
-        it('Make sure surveys are not loaded when config says no', () => {
-            given('decideResponse', () => ({
-                featureFlags: { 'test-flag': true },
-                surveys: true,
-            }))
-            given('config', () => ({
-                api_host: 'https://test.com',
-                token: 'testtoken',
-                persistence: 'memory',
-                disable_surveys: true,
-            }))
-
-            given.subject()
-            // Make sure the script is not loaded
-            expect(checkScriptsForSrc('https://test.com/static/surveys.js', true)).toBe(true)
+            expectScriptToNotExist('https://test.com/site_app/1/tokentoken/hash/')
         })
     })
 })