+++ src/__tests__/extensions/replay/sessionrecording.test.ts
@@ -216,7 +216,7 @@ describe('SessionRecording', () => {
         })
     })
 
-    describe('startRecordingIfEnabled', () => {
+    describe('startIfEnabledOrStop', () => {
         beforeEach(() => {
             // need to cast as any to mock private methods
             jest.spyOn(sessionRecording as any, '_startCapture')
@@ -225,12 +225,12 @@ describe('SessionRecording', () => {
         })
 
         it('call _startCapture if its enabled', () => {
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             expect((sessionRecording as any)._startCapture).toHaveBeenCalled()
         })
 
         it('emits an options event', () => {
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             expect((sessionRecording as any)['_tryAddCustomEvent']).toHaveBeenCalledWith('$session_options', {
                 activePlugins: [],
                 sessionRecordingOptions: {
@@ -253,18 +253,18 @@ describe('SessionRecording', () => {
 
         it('call stopRecording if its not enabled', () => {
             posthog.config.disable_session_recording = true
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             expect(sessionRecording.stopRecording).toHaveBeenCalled()
         })
     })
 
     describe('afterDecideResponse()', () => {
         beforeEach(() => {
-            jest.spyOn(sessionRecording, 'startRecordingIfEnabled')
+            jest.spyOn(sessionRecording, 'startIfEnabledOrStop')
         })
 
         it('buffers snapshots until decide is received and drops them if disabled', () => {
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             expect(loadScript).toHaveBeenCalled()
             expect(sessionRecording['status']).toBe('buffering')
             expect(sessionRecording['buffer']).toEqual(EMPTY_BUFFER)
@@ -285,7 +285,7 @@ describe('SessionRecording', () => {
         })
 
         it('emit is not active until decide is called', () => {
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             expect(loadScript).toHaveBeenCalled()
             expect(sessionRecording['status']).toBe('buffering')
 
@@ -294,7 +294,7 @@ describe('SessionRecording', () => {
         })
 
         it('sample rate is null when decide does not return it', () => {
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             expect(loadScript).toHaveBeenCalled()
             expect(sessionRecording['isSampled']).toBe(null)
 
@@ -355,7 +355,7 @@ describe('SessionRecording', () => {
                 })
             )
 
-            expect(sessionRecording.startRecordingIfEnabled).toHaveBeenCalled()
+            expect(sessionRecording.startIfEnabledOrStop).toHaveBeenCalled()
             expect(loadScript).toHaveBeenCalled()
             expect(posthog.get_property(SESSION_RECORDING_ENABLED_SERVER_SIDE)).toBe(true)
             expect(sessionRecording['_endpoint']).toEqual('/ses/')
@@ -365,7 +365,7 @@ describe('SessionRecording', () => {
     describe('recording', () => {
         describe('sampling', () => {
             it('does not emit to capture if the sample rate is 0', () => {
-                sessionRecording.startRecordingIfEnabled()
+                sessionRecording.startIfEnabledOrStop()
 
                 sessionRecording.afterDecideResponse(
                     makeDecideResponse({
@@ -380,7 +380,7 @@ describe('SessionRecording', () => {
             })
 
             it('does emit to capture if the sample rate is null', () => {
-                sessionRecording.startRecordingIfEnabled()
+                sessionRecording.startIfEnabledOrStop()
 
                 sessionRecording.afterDecideResponse(
                     makeDecideResponse({
@@ -392,7 +392,7 @@ describe('SessionRecording', () => {
             })
 
             it('stores excluded session when excluded', () => {
-                sessionRecording.startRecordingIfEnabled()
+                sessionRecording.startIfEnabledOrStop()
 
                 sessionRecording.afterDecideResponse(
                     makeDecideResponse({
@@ -404,7 +404,7 @@ describe('SessionRecording', () => {
             })
 
             it('does emit to capture if the sample rate is 1', () => {
-                sessionRecording.startRecordingIfEnabled()
+                sessionRecording.startIfEnabledOrStop()
 
                 _emit(createIncrementalSnapshot({ data: { source: 1 } }))
                 expect(posthog.capture).not.toHaveBeenCalled()
@@ -427,7 +427,7 @@ describe('SessionRecording', () => {
             })
 
             it('sets emit as expected when sample rate is 0.5', () => {
-                sessionRecording.startRecordingIfEnabled()
+                sessionRecording.startIfEnabledOrStop()
 
                 sessionRecording.afterDecideResponse(
                     makeDecideResponse({
@@ -465,7 +465,7 @@ describe('SessionRecording', () => {
                     },
                 })
 
-                sessionRecording.startRecordingIfEnabled()
+                sessionRecording.startIfEnabledOrStop()
 
                 sessionRecording['_onScriptLoaded']()
                 expect(assignableWindow.rrweb.record).toHaveBeenCalledWith(
@@ -481,7 +481,7 @@ describe('SessionRecording', () => {
             })
 
             it('skips when any config variable is missing', () => {
-                sessionRecording.startRecordingIfEnabled()
+                sessionRecording.startIfEnabledOrStop()
 
                 sessionRecording.afterDecideResponse(
                     makeDecideResponse({
@@ -501,7 +501,7 @@ describe('SessionRecording', () => {
         it('calls rrweb.record with the right options', () => {
             posthog.persistence?.register({ [CONSOLE_LOG_RECORDING_ENABLED_SERVER_SIDE]: false })
 
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             // maskAllInputs should change from default
             // someUnregisteredProp should not be present
             expect(assignableWindow.rrweb.record).toHaveBeenCalledWith({
@@ -523,7 +523,7 @@ describe('SessionRecording', () => {
         })
 
         it('records events emitted before and after starting recording', () => {
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             expect(loadScript).toHaveBeenCalled()
 
             _emit(createIncrementalSnapshot({ data: { source: 1 } }))
@@ -579,7 +579,7 @@ describe('SessionRecording', () => {
 
         it('buffers emitted events', () => {
             sessionRecording.afterDecideResponse(makeDecideResponse({ sessionRecording: { endpoint: '/s/' } }))
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             expect(loadScript).toHaveBeenCalled()
 
             _emit(createIncrementalSnapshot({ data: { source: 1 } }))
@@ -615,7 +615,7 @@ describe('SessionRecording', () => {
 
         it('flushes buffer if the size of the buffer hits the limit', () => {
             sessionRecording.afterDecideResponse(makeDecideResponse({ sessionRecording: { endpoint: '/s/' } }))
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             expect(loadScript).toHaveBeenCalled()
             const bigData = 'a'.repeat(RECORDING_MAX_EVENT_SIZE * 0.8)
 
@@ -634,7 +634,7 @@ describe('SessionRecording', () => {
         })
 
         it('maintains the buffer if the recording is buffering', () => {
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             expect(loadScript).toHaveBeenCalled()
 
             const bigData = 'a'.repeat(RECORDING_MAX_EVENT_SIZE * 0.8)
@@ -660,7 +660,7 @@ describe('SessionRecording', () => {
 
         it('flushes buffer if the session_id changes', () => {
             sessionRecording.afterDecideResponse(makeDecideResponse({ sessionRecording: { endpoint: '/s/' } }))
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
 
             expect(sessionRecording['buffer']?.sessionId).toEqual(null)
 
@@ -714,12 +714,12 @@ describe('SessionRecording', () => {
 
         it("doesn't load recording script if already loaded", () => {
             addRRwebToWindow()
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             expect(loadScript).not.toHaveBeenCalled()
         })
 
         it('loads recording script from right place', () => {
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
 
             expect(loadScript).toHaveBeenCalledWith('https://test.com/static/recorder.js?v=v0.0.1', expect.anything())
         })
@@ -727,7 +727,7 @@ describe('SessionRecording', () => {
         it('loads script after `_startCapture` if not previously loaded', () => {
             posthog.persistence?.register({ [SESSION_RECORDING_ENABLED_SERVER_SIDE]: false })
 
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             expect(loadScript).not.toHaveBeenCalled()
 
             sessionRecording['_startCapture']()
@@ -738,7 +738,7 @@ describe('SessionRecording', () => {
         it('does not load script if disable_session_recording passed', () => {
             posthog.config.disable_session_recording = true
 
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             sessionRecording['_startCapture']()
 
             expect(loadScript).not.toHaveBeenCalled()
@@ -747,7 +747,7 @@ describe('SessionRecording', () => {
         it('session recording can be turned on and off', () => {
             expect(sessionRecording['stopRrweb']).toEqual(undefined)
 
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
 
             expect(sessionRecording.started).toEqual(true)
             expect(sessionRecording['stopRrweb']).not.toEqual(undefined)
@@ -761,7 +761,7 @@ describe('SessionRecording', () => {
         it('session recording can be turned on after being turned off', () => {
             expect(sessionRecording['stopRrweb']).toEqual(undefined)
 
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
 
             expect(sessionRecording.started).toEqual(true)
             expect(sessionRecording['stopRrweb']).not.toEqual(undefined)
@@ -776,7 +776,7 @@ describe('SessionRecording', () => {
             it('if not enabled, plugin is not used', () => {
                 posthog.config.enable_recording_console_log = false
 
-                sessionRecording.startRecordingIfEnabled()
+                sessionRecording.startIfEnabledOrStop()
 
                 expect(assignableWindow.rrwebConsoleRecord.getRecordConsolePlugin).not.toHaveBeenCalled()
             })
@@ -784,7 +784,7 @@ describe('SessionRecording', () => {
             it('if enabled, plugin is used', () => {
                 posthog.config.enable_recording_console_log = true
 
-                sessionRecording.startRecordingIfEnabled()
+                sessionRecording.startIfEnabledOrStop()
 
                 expect(assignableWindow.rrwebConsoleRecord.getRecordConsolePlugin).toHaveBeenCalled()
             })
@@ -795,7 +795,7 @@ describe('SessionRecording', () => {
                 sessionRecording['sessionId'] = 'old-session-id'
                 sessionRecording['windowId'] = 'old-window-id'
 
-                sessionRecording.startRecordingIfEnabled()
+                sessionRecording.startIfEnabledOrStop()
                 sessionRecording.afterDecideResponse(
                     makeDecideResponse({
                         sessionRecording: { endpoint: '/s/' },
@@ -862,7 +862,7 @@ describe('SessionRecording', () => {
 
                     expect(mockCallback).not.toHaveBeenCalled()
 
-                    sessionRecording.startRecordingIfEnabled()
+                    sessionRecording.startIfEnabledOrStop()
                     sessionRecording['_startCapture']()
 
                     expect(mockCallback).toHaveBeenCalledTimes(1)
@@ -934,7 +934,7 @@ describe('SessionRecording', () => {
                     sessionManager = new SessionIdManager(config, new PostHogPersistence(config))
                     posthog.sessionManager = sessionManager
 
-                    sessionRecording.startRecordingIfEnabled()
+                    sessionRecording.startIfEnabledOrStop()
                     sessionRecording['_startCapture']()
                 })
 
@@ -1086,7 +1086,7 @@ describe('SessionRecording', () => {
         }
 
         beforeEach(() => {
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             sessionRecording.afterDecideResponse(makeDecideResponse({ sessionRecording: { endpoint: '/s/' } }))
             expect(sessionRecording['status']).toEqual('active')
 
@@ -1390,13 +1390,13 @@ describe('SessionRecording', () => {
 
     describe('buffering minimum duration', () => {
         it('can report no duration when no data', () => {
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             expect(sessionRecording['status']).toBe('buffering')
             expect(sessionRecording['sessionDuration']).toBe(null)
         })
 
         it('can report zero duration', () => {
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             expect(sessionRecording['status']).toBe('buffering')
             const { sessionStartTimestamp } = sessionManager.checkAndGetSessionAndWindowId(true)
             _emit(createIncrementalSnapshot({ data: { source: 1 }, timestamp: sessionStartTimestamp }))
@@ -1404,7 +1404,7 @@ describe('SessionRecording', () => {
         })
 
         it('can report a duration', () => {
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             expect(sessionRecording['status']).toBe('buffering')
             const { sessionStartTimestamp } = sessionManager.checkAndGetSessionAndWindowId(true)
             _emit(createIncrementalSnapshot({ data: { source: 1 }, timestamp: sessionStartTimestamp + 100 }))
@@ -1412,7 +1412,7 @@ describe('SessionRecording', () => {
         })
 
         it('starts with an undefined minimum duration', () => {
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             expect(sessionRecording['minimumDuration']).toBe(null)
         })
 
@@ -1431,7 +1431,7 @@ describe('SessionRecording', () => {
                     sessionRecording: { minimumDurationMilliseconds: 1500 },
                 })
             )
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             expect(sessionRecording['status']).toBe('active')
             const { sessionStartTimestamp } = sessionManager.checkAndGetSessionAndWindowId(true)
             _emit(createIncrementalSnapshot({ data: { source: 1 }, timestamp: sessionStartTimestamp + 100 }))
@@ -1451,7 +1451,7 @@ describe('SessionRecording', () => {
                     sessionRecording: { minimumDurationMilliseconds: 1500 },
                 })
             )
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             expect(sessionRecording['status']).toBe('active')
             const { sessionStartTimestamp } = sessionManager.checkAndGetSessionAndWindowId(true)
 
@@ -1476,7 +1476,7 @@ describe('SessionRecording', () => {
                     sessionRecording: { minimumDurationMilliseconds: 1500 },
                 })
             )
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             expect(sessionRecording['status']).toBe('active')
             const { sessionStartTimestamp } = sessionManager.checkAndGetSessionAndWindowId(true)
             _emit(createIncrementalSnapshot({ data: { source: 1 }, timestamp: sessionStartTimestamp + 100 }))
@@ -1517,7 +1517,7 @@ describe('SessionRecording', () => {
             })
 
             sessionRecording.afterDecideResponse(makeDecideResponse({ sessionRecording: { endpoint: '/s/' } }))
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             expect(loadScript).toHaveBeenCalled()
 
             expect(sessionRecording['queuedRRWebEvents']).toHaveLength(0)
@@ -1555,12 +1555,12 @@ describe('SessionRecording', () => {
         })
 
         it('schedules a snapshot on start', () => {
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             expect(sessionRecording['_fullSnapshotTimer']).not.toBe(undefined)
         })
 
         it('reschedules a snapshot, when we take a full snapshot', () => {
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             const startTimer = sessionRecording['_fullSnapshotTimer']
 
             _emit(createFullSnapshot())
@@ -1574,7 +1574,7 @@ describe('SessionRecording', () => {
         beforeEach(() => {
             jest.spyOn(sessionRecording as any, '_tryAddCustomEvent')
             posthog.config.capture_pageview = false
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             // clear the spy calls
             ;(sessionRecording as any)._tryAddCustomEvent.mockClear()
         })
@@ -1613,7 +1613,7 @@ describe('SessionRecording', () => {
         beforeEach(() => {
             jest.spyOn(sessionRecording as any, '_tryAddCustomEvent')
             posthog.config.capture_pageview = true
-            sessionRecording.startRecordingIfEnabled()
+            sessionRecording.startIfEnabledOrStop()
             // clear the spy calls
             ;(sessionRecording as any)._tryAddCustomEvent.mockClear()
         })