+++ src/posthog-core.ts
@@ -49,7 +49,7 @@ import {
     ToolbarParams,
 } from './types'
 import { SentryIntegration } from './extensions/sentry-integration'
-import { createSegmentIntegration } from './extensions/segment-integration'
+import { setupSegmentIntegration } from './extensions/segment-integration'
 import { PageViewManager } from './page-view'
 import { PostHogSurveys } from './posthog-surveys'
 import { RateLimiter } from './rate-limiter'
@@ -70,9 +70,10 @@ import { logger } from './utils/logger'
 import { SessionPropsManager } from './session-props'
 import { _isBlockedUA } from './utils/blocked-uas'
 import { extendURLParams, request, SUPPORTS_REQUEST } from './request'
-import { Autocapture } from './autocapture'
 import { Heatmaps } from './heatmaps'
 import { ScrollManager } from './scroll-manager'
+import { SimpleEventEmitter } from './utils/simple-event-emitter'
+import { Autocapture } from './autocapture'
 
 /*
 SIMPLE STYLE GUIDE:
@@ -122,7 +123,6 @@ export const defaultConfig = (): PostHogConfig => ({
     cross_subdomain_cookie: isCrossDomainCookie(document?.location),
     persistence: 'localStorage+cookie', // up to 1.92.0 this was 'cookie'. It's easy to migrate as 'localStorage+cookie' will migrate data from cookie storage
     persistence_name: '',
-    cookie_name: '',
     loaded: __NOOP,
     store_google: true,
     custom_campaign_params: [],
@@ -136,7 +136,6 @@ export const defaultConfig = (): PostHogConfig => ({
     upgrade: false,
     disable_session_recording: false,
     disable_persistence: false,
-    disable_cookie: false,
     disable_surveys: false,
     enable_recording_console_log: undefined, // When undefined, it falls back to the server-side setting
     secure_cookie: window?.location?.protocol === 'https:',
@@ -147,8 +146,6 @@ export const defaultConfig = (): PostHogConfig => ({
     opt_out_capturing_persistence_type: 'localStorage',
     opt_out_capturing_cookie_prefix: null,
     opt_in_site_apps: false,
-    // Deprecated, use property_denylist instead.
-    property_blacklist: [],
     property_denylist: [],
     respect_dnt: false,
     sanitize_properties: null,
@@ -177,8 +174,41 @@ export const defaultConfig = (): PostHogConfig => ({
     bootstrap: {},
     disable_compression: false,
     session_idle_timeout_seconds: 30 * 60, // 30 minutes
-    process_person: 'always',
+    person_profiles: 'always',
 })
+
+export const configRenames = (origConfig: Partial<PostHogConfig>): Partial<PostHogConfig> => {
+    const renames: Partial<PostHogConfig> = {}
+    if (!_isUndefined(origConfig.process_person)) {
+        renames.person_profiles = origConfig.process_person
+    }
+    if (!_isUndefined(origConfig.xhr_headers)) {
+        renames.request_headers = origConfig.xhr_headers
+    }
+    if (!_isUndefined(origConfig.cookie_name)) {
+        renames.persistence_name = origConfig.cookie_name
+    }
+    if (!_isUndefined(origConfig.disable_cookie)) {
+        renames.disable_persistence = origConfig.disable_cookie
+    }
+    // on_xhr_error is not present, as the type is different to on_request_error
+
+    // the original config takes priority over the renames
+    const newConfig = _extend({}, renames, origConfig)
+
+    // merge property_blacklist into property_denylist
+    if (_isArray(origConfig.property_blacklist)) {
+        if (_isUndefined(origConfig.property_denylist)) {
+            newConfig.property_denylist = origConfig.property_blacklist
+        } else if (_isArray(origConfig.property_denylist)) {
+            newConfig.property_denylist = [...origConfig.property_blacklist, ...origConfig.property_denylist]
+        } else {
+            logger.error('Invalid value for property_denylist config: ' + origConfig.property_denylist)
+        }
+    }
+
+    return newConfig
+}
 
 class DeprecatedWebPerformanceObserver {
     get _forceAllowLocalhost(): boolean {
@@ -226,13 +256,13 @@ export class PostHog {
 
     _triggered_notifs: any
     compression?: Compression
-    __captureHooks: ((eventName: string) => void)[]
     __request_queue: QueuedRequestOptions[]
     decideEndpointWasHit: boolean
     analyticsDefaultEndpoint: string
 
     SentryIntegration: typeof SentryIntegration
-    segmentIntegration: () => any
+
+    private _debugEventEmitter = new SimpleEventEmitter()
 
     /** DEPRECATED: We keep this to support existing usage but now one should just call .setPersonProperties */
     people: {
@@ -244,8 +274,6 @@ export class PostHog {
         this.config = defaultConfig()
         this.decideEndpointWasHit = false
         this.SentryIntegration = SentryIntegration
-        this.segmentIntegration = () => createSegmentIntegration(this)
-        this.__captureHooks = []
         this.__request_queue = []
         this.__loaded = false
         this.analyticsDefaultEndpoint = '/e/'
@@ -271,6 +299,8 @@ export class PostHog {
                 callback?.({} as any)
             },
         }
+
+        this.on('eventCaptured', (data) => logger.info('send', data))
     }
 
     // Initialization methods
@@ -340,10 +370,8 @@ export class PostHog {
         this.config = {} as PostHogConfig // will be set right below
         this._triggered_notifs = []
 
-        config.request_headers = config.request_headers || config.xhr_headers
-
         this.set_config(
-            _extend({}, defaultConfig(), config, {
+            _extend({}, defaultConfig(), configRenames(config), {
                 name: name,
                 token: token,
             })
@@ -359,18 +387,21 @@ export class PostHog {
 
         this._requestQueue = new RequestQueue((req) => this._send_request(req))
         this._retryQueue = new RetryQueue(this)
-        this.__captureHooks = []
         this.__request_queue = []
 
         this.sessionManager = new SessionIdManager(this.config, this.persistence)
         this.sessionPropsManager = new SessionPropsManager(this.sessionManager, this.persistence)
 
         this.sessionRecording = new SessionRecording(this)
-        this.sessionRecording.startRecordingIfEnabled()
+        this.sessionRecording.startIfEnabledOrStop()
 
         if (!this.config.disable_scroll_properties) {
             this.scrollManager.startMeasuringScrollPosition()
         }
+
+        this.autocapture = new Autocapture(this)
+        this.autocapture.startIfEnabled()
+        this.surveys.loadIfEnabled()
 
         this.autocapture = new Autocapture(this)
         this.heatmaps = new Heatmaps(this)
@@ -380,19 +411,6 @@ export class PostHog {
         Config.DEBUG = Config.DEBUG || this.config.debug
 
         this._gdpr_init()
-
-        if (config.segment) {
-            // Use segments anonymousId instead
-            this.config.get_device_id = () => config.segment.user().anonymousId()
-
-            // If a segment user ID exists, set it as the distinct_id
-            if (config.segment.user().id()) {
-                this.register({
-                    distinct_id: config.segment.user().id(),
-                })
-                this.persistence.set_user_state('identified')
-            }
-        }
 
         // isUndefined doesn't provide typehint here so wouldn't reduce bundle as we'd need to assign
         // eslint-disable-next-line posthog-js/no-direct-undefined-check
@@ -451,11 +469,13 @@ export class PostHog {
 
         // We wan't to avoid promises for IE11 compatibility, so we use callbacks here
         if (config.segment) {
-            config.segment.register(this.segmentIntegration()).then(() => {
-                this._loaded()
-            })
+            setupSegmentIntegration(this, () => this._loaded())
         } else {
             this._loaded()
+        }
+
+        if (_isFunction(this.config._onCapture)) {
+            this.on('eventCaptured', (data) => this.config._onCapture(data.event, data))
         }
 
         return this
@@ -475,6 +495,10 @@ export class PostHog {
         if (response.analytics?.endpoint) {
             this.analyticsDefaultEndpoint = response.analytics.endpoint
         }
+
+        this.sessionRecording?.afterDecideResponse(response)
+        this.autocapture?.afterDecideResponse(response)
+        this.surveys?.afterDecideResponse(response)
     }
 
     _loaded(): void {
@@ -770,7 +794,7 @@ export class PostHog {
             this.setPersonPropertiesForFlags(finalSet)
         }
 
-        logger.info('send', data)
+        this._debugEventEmitter.emit('eventCaptured', data)
 
         const requestOptions: QueuedRequestOptions = {
             method: 'POST',
@@ -786,18 +810,11 @@ export class PostHog {
             this._send_retriable_request(requestOptions)
         }
 
-        this._invokeCaptureHooks(event_name, data)
-
         return data
     }
 
     _addCaptureHook(callback: (eventName: string) => void): void {
-        this.__captureHooks.push(callback)
-    }
-
-    _invokeCaptureHooks(eventName: string, eventData: CaptureResult): void {
-        this.config._onCapture(eventName, eventData)
-        _each(this.__captureHooks, (callback) => callback(eventName))
+        this.on('eventCaptured', (data) => callback(data.event))
     }
 
     _calculate_event_properties(event_name: string, event_properties: Properties): Properties {
@@ -880,11 +897,8 @@ export class PostHog {
 
         properties['$is_identified'] = this._isIdentified()
 
-        if (_isArray(this.config.property_denylist) && _isArray(this.config.property_blacklist)) {
-            // since property_blacklist is deprecated in favor of property_denylist, we merge both of them here
-            // TODO: merge this only once, requires refactoring tests
-            const property_denylist = [...this.config.property_blacklist, ...this.config.property_denylist]
-            _each(property_denylist, function (denylisted_prop) {
+        if (_isArray(this.config.property_denylist)) {
+            _each(this.config.property_denylist, function (denylisted_prop) {
                 delete properties[denylisted_prop]
             })
         } else {
@@ -902,7 +916,7 @@ export class PostHog {
         }
 
         // add person processing flag as very last step, so it cannot be overridden. process_person=true is default
-        properties['$process_person'] = this._hasPersonProcessing()
+        properties['$process_person_profile'] = this._hasPersonProcessing()
 
         return properties
     }
@@ -1082,6 +1096,18 @@ export class PostHog {
     /** Get the list of early access features. To check enrollment status, use `isFeatureEnabled`. */
     getEarlyAccessFeatures(callback: EarlyAccessFeatureCallback, force_reload = false): void {
         return this.featureFlags.getEarlyAccessFeatures(callback, force_reload)
+    }
+
+    /**
+     * Exposes a set of events that PostHog will emit.
+     * e.g. `eventCaptured` is emitted immediately before trying to send an event
+     *
+     * Unlike  `onFeatureFlags` and `onSessionId` these are not called when the
+     * listener is registered, the first callback will be the next event
+     * _after_ registering a listener
+     */
+    on(event: 'eventCaptured', cb: (...args: any[]) => void): () => void {
+        return this._debugEventEmitter.on(event, cb)
     }
 
     /*
@@ -1643,14 +1669,7 @@ export class PostHog {
     set_config(config: Partial<PostHogConfig>): void {
         const oldConfig = { ...this.config }
         if (_isObject(config)) {
-            _extend(this.config, config)
-
-            if (!this.config.persistence_name) {
-                this.config.persistence_name = this.config.cookie_name
-            }
-            if (!this.config.disable_persistence) {
-                this.config.disable_persistence = this.config.disable_cookie
-            }
+            _extend(this.config, configRenames(config))
 
             this.persistence?.update_config(this.config, oldConfig)
             this.sessionPersistence =
@@ -1665,21 +1684,9 @@ export class PostHog {
                 Config.DEBUG = true
             }
 
-            if (this.sessionRecording && !_isUndefined(config.disable_session_recording)) {
-                const disable_session_recording_has_changed =
-                    oldConfig.disable_session_recording !== config.disable_session_recording
-                // if opting back in, this config might not have changed
-                const try_enable_after_opt_in =
-                    !userOptedOut(this) && !config.disable_session_recording && !this.sessionRecording.started
-
-                if (disable_session_recording_has_changed || try_enable_after_opt_in) {
-                    if (config.disable_session_recording) {
-                        this.sessionRecording.stopRecording()
-                    } else {
-                        this.sessionRecording.startRecordingIfEnabled()
-                    }
-                }
-            }
+            this.sessionRecording?.startIfEnabledOrStop()
+            this.autocapture?.startIfEnabled()
+            this.surveys.loadIfEnabled()
         }
     }
 
@@ -1688,7 +1695,7 @@ export class PostHog {
      * disable_session_recording to false
      * @param override.sampling - optional boolean to override the default sampling behavior - ensures the next session recording to start will not be skipped by sampling config.
      */
-    startSessionRecording(override: { sampling?: boolean }): void {
+    startSessionRecording(override?: { sampling?: boolean }): void {
         if (override?.sampling) {
             // allow the session id check to rotate session id if necessary
             const ids = this.sessionManager?.checkAndGetSessionAndWindowId()
@@ -1788,8 +1795,8 @@ export class PostHog {
 
     _hasPersonProcessing(): boolean {
         return !(
-            this.config.process_person === 'never' ||
-            (this.config.process_person === 'identified_only' &&
+            this.config.person_profiles === 'never' ||
+            (this.config.person_profiles === 'identified_only' &&
                 !this._isIdentified() &&
                 _isEmptyObject(this.getGroups()) &&
                 !this.persistence?.props?.[ALIAS_ID_KEY] &&
@@ -1803,7 +1810,7 @@ export class PostHog {
      * @param function_name
      */
     _requirePersonProcessing(function_name: string): boolean {
-        if (this.config.process_person === 'never') {
+        if (this.config.person_profiles === 'never') {
             logger.error(
                 function_name + ' was called, but process_person is set to "never". This call will be ignored.'
             )