+++ CHANGELOG.md
@@ -1,3 +1,53 @@
+## 1.126.0 - 2024-04-17
+
+- fix: Reacting to config changes (#1138)
+
+## 1.125.0 - 2024-04-16
+
+- Rename $process_person to $process_person_profile (#1143)
+
+## 1.124.0 - 2024-04-16
+
+- Update docs on versioning (#1142)
+- feat: Update support for segment analytics (#1119)
+
+## 1.123.1 - 2024-04-16
+
+- feat: Update support for segment analytics (#1119)
+
+## 1.123.0 - 2024-04-16
+
+- feat: Rename process_person config option to person_profiles (#1141)
+- feat: readonly event emitter (#1136)
+- fix: Allow capturing `aria-label` attribute on sensitive elements (#1118)
+- chore: Autocapture refactor (#1126)
+- chore: patch to fix broken worker (#1134)
+- fix: continue recording after reset (#1135)
+- fix: sampling typescript error (#1129)
+- chore: no only tests are allowed (#1128)
+
+## 1.122.0 - 2024-04-15
+
+- feat: readonly event emitter (#1136)
+- fix: Allow capturing `aria-label` attribute on sensitive elements (#1118)
+
+## 1.121.4 - 2024-04-15
+
+- chore: Autocapture refactor (#1126)
+
+## 1.121.3 - 2024-04-15
+
+- chore: patch to fix broken worker (#1134)
+
+## 1.121.2 - 2024-04-12
+
+- fix: continue recording after reset (#1135)
+
+## 1.121.1 - 2024-04-10
+
+- fix: sampling typescript error (#1129)
+- chore: no only tests are allowed (#1128)
+
 ## 1.121.0 - 2024-04-10
 
 - feat(surveys): Add disable config option (#1123)